move evaluators to independent module
